Project-Expense tracking(python)/requirement.txt
pandas==2.2.3
streamlit==1.43.1
fastapi==0.115.11
pydantic==2.10.6
uvicorn==0.34.0
mysql-connector-python==9.2.0
requests==2.32.3
pytest==8.3.5